fix: dropdown close function check if domRef.current (#202)
* fix: dropdown close function check if domRef.current

In Edge 18, if the dropdown is toggled by an external flag controlled by
a click event, the domRef becomes invalid before the event listener is
removed, which leads to this.dom.Ref.current.contains to crash. Simply
checking for this will result in this.setState being called in an
unmounted component.

This PR checks if domRef.current exists and only calls this.setState if
the domRef is still valid. It also uses mount to test the dropdown state
changes.

* Update dropdown.js